requirements.txt
# Core & Scheduling
apscheduler==3.10.4
python-dotenv==1.0.0

# Data Extraction
requests==2.31.0
feedparser==6.0.10
sec-api==0.0.1

# Analysis
semantic-kernel==1.34.0
openai==1.67.0

# Validation
google-api-python-client==2.108.0

# Persistence & Reporting
pandas==2.1.3
tabulate==0.9.0 